+++ keybert/__init__.py
@@ -1,3 +1,4 @@
+from keybert._llm import KeyLLM
 from keybert._model import KeyBERT
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"